% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleSeries.R
\name{AlleleSeries}
\alias{AlleleSeries}
\title{Form QTL incidence matrices}
\usage{
AlleleSeries(na, a)
}
\arguments{
\item{na}{Number of alleles}

\item{a}{parameter of the function}
}
\value{
vector with values of the allelic series
}
\description{
The function for the QTL incidence matrices of position given different type
of effects: cross-specific, parental, ancestral and bi-allelic.
}
