% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cross_scheme_NAM.R
\name{cross_scheme_NAM}
\alias{cross_scheme_NAM}
\title{cross_scheme_NAM}
\usage{
cross_scheme_NAM(np)
}
\arguments{
\item{np}{Parent number in the design.}
}
\value{
parent per cross information
}
\description{
Form crossing scheme for a NAM population with np parents.
}
\author{
Vincent Garin
}
