\name{EUNAM_map}
\alias{EUNAM_map}
\docType{data}
\title{
  Reduced map maize EU-NAM Dent population
}
\description{
  Reduced genetic map of the maize EU nested association mapping (EUNAM) Dent population
  (Bauer et al., 2013).
}
\usage{data(EUNAM_map)}
\format{
  \code{data.frame}
}
\details{
  Selection of 1000 markers from the genetic map of the Maize EU-NAM
  population.
}
\source{
  \url{http://maizegdb.org/data_center/reference?id=9024747}
}
\references{
  Bauer, E., Falque, M., Walter, H., Bauland, C., Camisan, C., Campo, L., ... & Altmann, T. (2013).
  Intraspecific variation of recombination rate in maize. Genome biology, 14(9), R103.

}
\examples{
  data(EUNAM_map)
}


\keyword{datasets}
