% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR2.R
\name{FDR2}
\alias{FDR2}
\title{False discovery rate 2}
\usage{
FDR2(QTL_true, QTL_detected, n_des = 9, n_mod = 4, n_chr)
}
\arguments{
\item{QTL_true}{list of true QTL positions.}

\item{QTL_detected}{list results of detected QTLs.}

\item{n_des}{Number of MPP design. Default = 9.}

\item{n_mod}{Number of QTL detection models. Default = 4.}

\item{n_chr}{Number of chromosome.}
}
\value{
FDR in percent
}
\description{
Alternative way to measure the false discovery rate looking at the results on
the chromosome where no signal was simulated.
}
