% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDRRegTable.R
\name{FDRRegTable}
\alias{FDRRegTable}
\title{FDR regression table results}
\usage{
FDRRegTable(QTL_true, QTL_detected, d_QTL, n_des = 9, n_mod = 4,
  mod_names = c("cr", "par", "anc", "biall"), des_names, n_par,
  N = 800)
}
\arguments{
\item{QTL_true}{list of true QTL positions.}

\item{QTL_detected}{list results of detected QTLs.}

\item{d_QTL}{distance to the QTL.}

\item{n_des}{Number of MPP design. Default = 9.}

\item{n_mod}{Number of QTL detection models. Default = 4.}

\item{mod_names}{Models names. Default = c("cr", "par", "anc", "biall").}

\item{des_names}{MPP design names. Use only: 'Diallel', 'Chessboard',
'Factorial', 'NAM' or 'Realized'. Default = NULL.}

\item{n_par}{Number of parents of the different designs.}

\item{N}{Total number of individuals in the population. Default = 800.}
}
\value{
...
}
\description{
FDR regression table results
}
\author{
Vincent Garin
}
