% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDRTable.R
\name{FDRTable}
\alias{FDRTable}
\title{FDR Table}
\usage{
FDRTable(
  QTL_true,
  QTL_detected,
  d_QTL,
  n_des = 9,
  n_mod = 4,
  n_QTL = 2,
  mod_names = c("cr", "par", "anc", "biall"),
  des_names = NULL,
  n_par,
  N
)
}
\arguments{
\item{QTL_true}{list of true QTL positions.}

\item{QTL_detected}{list results of detected QTLs.}

\item{d_QTL}{distance to the QTL.}

\item{n_des}{Number of MPP design. Default = 9.}

\item{n_mod}{Number of QTL detection models. Default = 4.}

\item{n_QTL}{Number of QTL. Default = 2.}

\item{mod_names}{Models names. Default = c("cr", "par", "anc", "biall").}

\item{des_names}{MPP design names.}
}
\value{
Average TPR per MPP design and QTL detection model.
}
\description{
FDR Table
}
\author{
Vincent Garin
}
