% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_par_per_cross.R
\name{form_par_per_cross}
\alias{form_par_per_cross}
\title{form parent per cross information in a diallel}
\usage{
form_par_per_cross(np = 9)
}
\arguments{
\item{np}{Parent number in the diallel.}
}
\value{
parent per cross information
}
\description{
form parent per cross information in a diallel
}
\author{
Vincent Garin
}
