% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formMppData.R
\name{formMppData}
\alias{formMppData}
\title{Form mppData objects}
\usage{
formMppData(
  geno_ABH,
  geno_012,
  map,
  MPP_sub,
  mk_prob,
  pheno,
  type = "F",
  nb_gen = 2
)
}
\arguments{
\item{geno_ABH}{genotype matrix of the global population in ABH format.}

\item{geno_012}{genotype matrix of the global population in 012 format.}

\item{map}{Correspondig three column genetic map (mk.id, chr, pos cM).}

\item{MPP_sub}{subset information obtained as output of the function
\code{\link{SampleMPPDesign}}.}

\item{mk_prob}{Problematic markers in the subset of the global population}

\item{pheno}{Simulated phenotype values.}

\item{type}{Type of population. Default = "F"}

\item{nb_gen}{Number of generation. Default = 2.}
}
\value{
list containing two type of mppData object
}
\description{
Function that subset a global population composed from all the cross of a
Dialleld structure between 9 parents to for mppData objects. The function
start from a particular realization (global population) composed of 5000
markers.
}
\author{
Vincent Garin
}
