\name{geno_par}
\alias{geno_par}
\docType{data}
\title{
  Selection parent genotype data EU-NAM Dent population
}
\description{
  Selection of parent genotype data from maize the EU-NAM Dent population
  (Bauer et al., 2013).
}
\usage{data(geno_par)}
\format{
  \code{matrix}
}
\details{
  1000 markers sample of the parent marker matrix of the EU-NAM Dent population.
}
\source{
  \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE50558}
}
\references{

  Bauer, E., Falque, M., Walter, H., Bauland, C., Camisan, C., Campo, L., ... & Altmann, T. (2013). Intraspecific variation of recombination rate in maize. Genome biology, 14(9), R103.

}
\examples{
  data(geno_par)
}


\keyword{datasets}
