% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncMatQTL.R
\name{IncMatQTL}
\alias{IncMatQTL}
\title{Form QTL incidence matrices}
\usage{
IncMatQTL(QTL, mppData, mppData_bi, par_clu)
}
\arguments{
\item{QTL}{list of QTL position obtained with function \code{\link{QposSelect}}}

\item{mppData}{IBD \code{mppData} object}

\item{mppData_bi}{IBS \code{mppData} object}

\item{par_clu}{parent clustering object.}
}
\value{
list of QTL incidence matrices
}
\description{
The function for the QTL incidence matrices of position given different type
of effects: cross-specific, parental, ancestral and bi-allelic.
}
