% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_LD_computation.R
\name{LD_computation}
\alias{LD_computation}
\title{LD_computation}
\usage{
LD_computation(
  geno,
  map,
  pop_ind = NULL,
  max_g_dist = 30,
  chr_sel = 1,
  prop_mk_comb = 1
)
}
\arguments{
\item{geno}{\code{Numeric} genotype marker matrix coded in 0, 1, 2 representing
the number of allele copies of one allele (e.g. second parent, minor allele).
Genotypes in row and markers in column. The column (marker) name should be
identical to the first column of the map (marker indicators).
Can be obtrained with \code{\link{sim_mpp_cross}}}

\item{map}{\code{data.frame} representing the marker map information with
the following columns: 1) \code{character} marker identifier;
2) \code{numeric} chromosomes (1, 2, 3, ...);
3) \code{numeric} genetic distance in cM;
4) \code{numeric} Optional physical distance in bp.}

\item{pop_ind}{Optional population indicator \code{Character vector}
indicating to which (sup)population each genotype belongs. Default = NULL}

\item{max_g_dist}{\code{Numeric value} between 0 and 1 indicating the
proportion of marker pairs selected to calculate the LD. Default = 1}
}
\value{
\code{data.frame} with marker i and j position, genetic distance
(physical distance) beteen markers and LD
}
\description{
Calculate linkage disequilibrium (LD) between a set of markers
}
\examples{

data("geno_par")
rownames(geno_par) <- paste0('P', 1:nrow(geno_par))

data("EUNAM_map")
rownames(map) <- map[, 1]

# NAM crossing scheme with 3 parents
cross_scheme <- cross_scheme_NAM(3)

geno <- sim_mpp_cross(geno_par = geno_par, map = map,
                     cross_scheme = cross_scheme,
                     n_ind_cr = c(100, 100),
                     geno_score_format = "numeric_IBD")

cr_ind <- substr(x = rownames(geno), start = 1, stop = 3)

LD_chr1 <- LD_computation(geno = geno, map = map, pop_ind = cr_ind,
                         max_g_dist = 40, chr_sel = 1, prop_mk_comb = 1)


}
