% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_LD_plot.R
\name{LD_plot}
\alias{LD_plot}
\title{LD_plot}
\usage{
LD_plot(LD_data)
}
\arguments{
\item{LD_data}{\code{data.frame} obtained from function \code{\link{LD_computation}}}
}
\value{
plot
}
\description{
Linkage disequilibium decay plot
}
\examples{

data("geno_par")
rownames(geno_par) <- paste0('P', 1:nrow(geno_par))

data("EUNAM_map")
rownames(map) <- map[, 1]

# NAM crossing scheme with 3 parents
cross_scheme <- cross_scheme_NAM(3)

geno <- sim_mpp_cross(geno_par = geno_par, map = map,
                     cross_scheme = cross_scheme,
                     n_ind_cr = c(100, 100),
                     geno_score_format = "numeric_IBD")

cr_ind <- substr(x = rownames(geno), start = 1, stop = 3)

LD_chr1 <- LD_computation(geno = geno, map = map, pop_ind = cr_ind,
                         max_g_dist = 40, chr_sel = 1, prop_mk_comb = 1)

LD_plot(LD_chr1)

}
