% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NindCr.R
\name{NindCr}
\alias{NindCr}
\title{Nb individual per cross}
\usage{
NindCr(N, Npar, MPP_des)
}
\arguments{
\item{N}{Total number of individual in the population}

\item{Npar}{Number of parents}

\item{MPP_des}{String indicating the type of MPP design among: 'Diallel',
'Chessboard', 'Factorial', 'NAM', 'Realized'}
}
\value{
Number of individuals per cross
}
\description{
Function to determine the average number of individual per cross given the
total size of the population, the type of MPP design and the number of parents
}
\author{
Vincent Garin
}
