% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TP_FP.R
\name{plot_TP_FP}
\alias{plot_TP_FP}
\title{QTL profile plot with true and false positive QTLs}
\usage{
plot_TP_FP(Qprof, QTL_TRUE, QTL_TP = NULL, QTL_FP = NULL, type = "l",
  main = "QTL profile", threshold = 3, text.size = 18)
}
\arguments{
\item{Qprof}{Object of class \code{QTLprof} returned by the function
\code{\link{mpp_SIM}} or \code{\link{mpp_CIM}}, or three columns
\code{numeric matrix} with chromosome, marker position and -log10(p-values).}

\item{QTL_TRUE}{Optional argument. List of QTL true QTL positions}

\item{QTL_TP}{Optional argument. List of false positive QTLs.}

\item{QTL_FP}{Optional argument. List of QTL true positive QTLs.}

\item{type}{\code{Character} expression indicating the type of plot should be
drawn: "l" for lines , "h" for vertical bar. Default = "l".}

\item{main}{Title of the graph. Default = "QTL profile".}

\item{threshold}{\code{Numeric} QTL significance threshold value draw on
the plot. Default = 3.}

\item{text.size}{\code{Numeric} value specifying the size of graph axis text
elements. Default = 18.}
}
\description{
QTL profile plot with true and false positive QTLs
}
\examples{

Not now

}
\author{
Vincent Garin
}
