% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMarkers.R
\name{QCMarkers}
\alias{QCMarkers}
\title{Quality control on markers}
\usage{
QCMarkers(geno, cross_ind, parallel = FALSE, cluster = NULL)
}
\arguments{
\item{geno}{genotype marker matrix}

\item{cross_ind}{Cross indicator}
}
\value{
Numeric vector with problematic markers.
}
\description{
Perform a quick QC on marker to detect the one that do not segregate in at
least on of the cross with 25%. The monomorphic markers are therefore also
tagged.
}
\author{
Vincent Garin
}
