% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QposSelect.R
\name{QposSelect}
\alias{QposSelect}
\title{Select QTL positions}
\usage{
QposSelect(N = 8, map, dist = 30)
}
\arguments{
\item{N}{Number of positions. Default = 8}

\item{map}{Three columns map with marker id, chromosome and position in cM.}

\item{dist}{Minimum distance between the positions on a same chromosome.
Default = 30.}
}
\value{
Three columns \code{data.frame} with the selected positions
}
\description{
The function select QTL positions that will be used for simulation. First
the positions are peaked on different chromosome. Then, the next positions
are chosen with a minimum distance in between.
}
\examples{

# Later

}
\author{
Vincent Garin
}
