% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_FDR.R
\name{QTL_FDR}
\alias{QTL_FDR}
\title{Details QTL FDR}
\usage{
QTL_FDR(QTL.true, QTL.detected, distance = 5)
}
\arguments{
\item{QTL.true}{\code{data.frame} of true QTL positions with marker id,
chromosome and position in cM.}

\item{QTL.detected}{\code{data.frame} of detected QTL positions with marker
id, chromosome and position in cM.}

\item{distance}{maximal distance between the true QTL position and the detected
signal. Default = 5.}
}
\value{
FDR
}
\description{
Function that return the list of FDR QTL with the distance to the nearest
simlated QTL. If no QTL was simulated on the chromosome return 999.
}
\author{
Vincent Garin
}
