% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_TPR.R
\name{QTL_TPR}
\alias{QTL_TPR}
\title{Details QTL TPR}
\usage{
QTL_TPR(QTL.true, QTL.detected, distance = 5)
}
\arguments{
\item{QTL.true}{\code{data.frame} of true QTL positions with marker id,
chromosome and position in cM.}

\item{QTL.detected}{\code{data.frame} of detected QTL positions with marker
id, chromosome and position in cM.}

\item{distance}{maximal distance between the true QTL position and the detected
signal. Default = 5.}
}
\value{
FDR
}
\description{
Function that return the list of TPR QTL with the -log10(p-value) and the
distance to the simulated QTL.
}
\author{
Vincent Garin
}
