% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLDetection2.R
\name{QTLDetection2}
\alias{QTLDetection2}
\title{QTL detection}
\usage{
QTLDetection2(mppData, thre_cr, thre_par, thre_anc, thre_biall,
  win.cof = 50, win.QTL = 30, folder, Rep_id, n.cores = 1)
}
\arguments{
\item{mppData}{\code{mppData} object}

\item{thre_cr}{Significance thresold for the cross-specific model.}

\item{thre_par}{Significance threshold for the parental model.}

\item{thre_anc}{Significance threshold for the ancestral model.}

\item{thre_biall}{Significance threshold for the bi-allelic model.}

\item{win.cof}{Minimum distance between two cofactors. Default = 50.}

\item{win.QTL}{Minimum distance between two selected QTLs. Default = 30.}

\item{folder}{Folder location where the results are saved.}

\item{Rep_id}{Identifier for the specific replication.}

\item{n.cores}{number of cores to use for parallelization. Default = 1}
}
\value{
list with the QTLs detected with the four models.
}
\description{
Function to perform QTL detection using four types of models: cross-specific,
parental, ancestral and bi-allelic. The function save the results of the
QTL detection processes in a specified folder and return the detected QTLs.
}
\author{
Vincent Garin
}
