% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDQ.R
\name{RDQ}
\alias{RDQ}
\title{Resolution distance to QTL}
\usage{
RDQ(QTL.true, QTL.detected, distance = 10)
}
\arguments{
\item{QTL.true}{\code{data.frame} of true QTL positions with marker id,
chromosome and position in cM.}

\item{QTL.detected}{\code{data.frame} of detected QTL positions with marker
id, chromosome and position in cM.}

\item{distance}{maximal distance between the true QTL position and the detected
signal. Default = 10.}
}
\value{
Vector with distance to the position. NA if nothing was detected in
the surrounding of the QTL.
}
\description{
Function measuring the distance between a "true" QTL position and a detected
QTL. The function looks if there is a signal at +- distance of the QTL. If a
signal was detected in this region, it measures the distance between the QTL
position and the detected position.
}
\author{
Vincent Garin
}
