% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMPPDesign2.R
\name{SampleMPPDesign}
\alias{SampleMPPDesign}
\title{Sample MPP design}
\usage{
SampleMPPDesign(
  np_dia = 9,
  n_ind = 200,
  MPP_des,
  np_sel,
  N = 800,
  cr_list,
  ncr
)
}
\arguments{
\item{np_dia}{Parent number in the reference diallel. Default = 9.}

\item{n_ind}{Number of individual per cross in the reference diallel.
Default = 200.}

\item{MPP_des}{Selected MPP design for the MPP sampling. One of: 'NAM' for
nested association mapping design; 'Dia' for half-diallel; 'Chess' for
chessboard design; 'Fact' for factorial desian; and 'Real' for a realized
design.}

\item{np_sel}{Number of parents in the selected MPP design}

\item{N}{Total number of sampled individuals. Default = 800.}

\item{cr_list}{Numeric indicator of the desired crosses on the half diallel
grid. This argument is only necessary for 'free' MPP design.}

\item{ncr}{Numeric vector specifying the number of individuals that should be
selected in each cross in the free MPP design.}
}
\value{
List containing the numeric genotype identifiers, the selected parents
and the selected crosses.
}
\description{
Function to sample the realisation of a MPP design from a full diallel made
from \code{np_dia} with n_ind individual per crosses. The function sample a
realisation of the design specified in argument \code{MPP_des} composed from
\code{np_sel} parents including \code{N} individuals.
}
\author{
Vincent Garin
}
