% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_star.R
\name{sign_star}
\alias{sign_star}
\title{Significance star}
\usage{
sign_star(x)
}
\arguments{
\item{x}{significance value between 0 and 1.}
}
\value{
Star value
}
\description{
Function that convert significance of results into stars
}
