% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mpp_cross.R
\name{sim_mpp_cross}
\alias{sim_mpp_cross}
\title{Simulate MPP crosses population}
\usage{
sim_mpp_cross(geno_par, map, cross_scheme, n_ind_cr)
}
\arguments{
\item{geno_par}{Character genotype marker matrix of the parents used for the simulation.
The allele must be coded with one letter per allele. For example, AA, CC, GG,
etc. The rownames must specify the parent identifiers used in the crossing
scheme (cross_scheme).}

\item{map}{Data.frame. Correspondig three columns genetic map
(marker, chr, pos in cM).}

\item{cross_scheme}{Three column marker matrix specifying: the cross
indicator, the first parent (male), and the second parent (female).
The parents identifiers must be the same as the rownames of geno_par.}

\item{n_ind_cr}{numeric vector indicating the number of simulated genotype
per cross.}
}
\value{
Genotype marker matrix of all simulated crosses
}
\description{
The function simulates a multi-parent population composed of N F2 crosses.
It uses the genotypes data provided in the argument geno_par. The cross
scheme is specified in the argument cross_scheme. The functions call
function from the simcross package (Broman, 2019)
}
\references{
Karl W. Broman (2019) R/simcross: an R package for simulating and plotting general
experimental crosses. \url{https://github.com/kbroman/simcross}
}
\author{
Vincent Garin
}
