% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPorFP.R
\name{TPorFP}
\alias{TPorFP}
\title{TP or FP QTL}
\usage{
TPorFP(QTL.true, QTL.detected, distance = 5)
}
\arguments{
\item{QTL.true}{\code{data.frame} of true QTL positions with marker id,
chromosome and position in cM.}

\item{QTL.detected}{\code{data.frame} of detected QTL positions with marker
id, chromosome and position in cM.}

\item{distance}{maximal distance between the true QTL position and the detected
signal. Default = 5.}
}
\value{
A logical vector: TRUE for TP QTLs and FALSE for FP QTLs
}
\description{
Given a list of TRUE QTL and a list of detected QTL, the function return
for each QTL if it was a true positive or a false positive.
}
\author{
Vincent Garin
}
