% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPR.R
\name{TPR}
\alias{TPR}
\title{True positive rate}
\usage{
TPR(QTL.true, QTL.detected, distance = 5)
}
\arguments{
\item{QTL.true}{\code{data.frame} of true QTL positions with marker id,
chromosome and position in cM.}

\item{QTL.detected}{\code{data.frame} of detected QTL positions with marker
id, chromosome and position in cM.}

\item{distance}{maximal distance between the true QTL position and the detected
signal. Default = 5.}
}
\value{
TPR
}
\description{
Function computing the true positive rate (TPR) as the percentage of true
QTL positions that have been detected at a certain distance
}
\author{
Vincent Garin
}
