% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPR_QTL_detail.R
\name{TPR_QTL_detail}
\alias{TPR_QTL_detail}
\title{Detail results TPR QTLs}
\usage{
TPR_QTL_detail(QTL_true, QTL_detected, d_QTL, n_des = 9, n_mod = 4,
  mod_names = c("cr", "par", "anc", "biall"), des_names, n_par, N)
}
\arguments{
\item{QTL_true}{list of true QTL positions.}

\item{QTL_detected}{list results of detected QTLs.}

\item{d_QTL}{distance to the QTL.}

\item{n_des}{Number of MPP design. Default = 9.}

\item{n_mod}{Number of QTL detection models. Default = 4.}

\item{mod_names}{Models names. Default = c("cr", "par", "anc", "biall").}

\item{des_names}{MPP design names. Use only: 'Diallel', 'Chessboard',
'Factorial', 'NAM' or 'Realized'. Default = NULL.}

\item{n_par}{Number of parents of the different designs}

\item{N}{Total number of individuals in the population}
}
\value{
A data.frame with for each TPR QTL the replication, the design,
the number of parents of the design, the number of indidual per cross,
the model, the -log10(p.value) and the distance to a simulated QTL,
(999 if the QTL was on a chromosome with no simulated QTL). For the
run (combination of rep, design and model) where no TPR QTL was present,
the -log10pval and the distance to QTL are set to NA.
}
\description{
Detail results TPR QTLs
}
\author{
Vincent Garin
}
