% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPRmodelQTL.R
\name{TPRmodelQTL}
\alias{TPRmodelQTL}
\title{TPR for model x QTL combination}
\usage{
TPRmodelQTL(
  QTL_true,
  QTL_detected,
  d_QTL,
  n_des = 9,
  n_mod = 4,
  n_QTL = 8,
  mod_names = c("cr", "par", "anc", "biall")
)
}
\arguments{
\item{QTL_true}{list of true QTL positions.}

\item{QTL_detected}{list results of detected QTLs.}

\item{d_QTL}{distance to the QTL.}

\item{n_des}{Number of MPP design. Default = 9.}

\item{n_mod}{Number of QTL detection models. Default = 4.}

\item{n_QTL}{Number of QTL. Default = 8.}

\item{mod_names}{Models names. Default = c("cr", "par", "anc", "biall").}
}
\value{
A matrix with the precentage of detection over the whole simulation
for each type of QTL and model combination.
}
\description{
Computes and organises the TPR results for each combination of model and type
of QTL. The TPR is averaged over the repetition and MPP design. The function
returns the percentage of time a specific type
of QTL effect was detected by a specific model over a number of repetitions
and MPP design sampling.
}
\author{
Vincent Garin
}
