% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPRQTLmodelMPP.R
\name{TPRQTLmodelMPP}
\alias{TPRQTLmodelMPP}
\title{TPR for QTL x model x MPP combination}
\usage{
TPRQTLmodelMPP(
  QTL_true,
  QTL_detected,
  d_QTL,
  n_des = 9,
  n_mod = 4,
  n_QTL = 8,
  mod_names = c("cr", "par", "anc", "biall"),
  MPP_names = NULL
)
}
\arguments{
\item{QTL_true}{list of true QTL positions.}

\item{QTL_detected}{list results of detected QTLs.}

\item{d_QTL}{distance to the QTL.}

\item{n_des}{Number of MPP design. Default = 9.}

\item{n_mod}{Number of QTL detection models. Default = 4.}

\item{n_QTL}{Number of QTL. Default = 8.}

\item{mod_names}{Models names. Default = c("cr", "par", "anc", "biall").}

\item{MPP_names}{MPP design names.}
}
\value{
A matrix with the precentage of detection over the repetition of the
simulation procedure.
}
\description{
Computes and organises the TPR results for each combination of QTL, model,
and MPP design. The TPR is averaged over the repetitions. The function
returns the average and the standard deviation of the TPR for a specific
combination of QTL model and MPP design.
}
\author{
Vincent Garin
}
