% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPRresProcessing.R
\name{TPRresProcessing}
\alias{TPRresProcessing}
\title{Processing TPR results}
\usage{
TPRresProcessing(
  QTL_true,
  QTL_detected,
  d_QTL,
  n_des = 9,
  n_mod = 4,
  mod_names = c("cr", "par", "anc", "biall"),
  MPP_names
)
}
\arguments{
\item{QTL_true}{list of true QTL positions.}

\item{QTL_detected}{list results of detected QTLs.}

\item{d_QTL}{distance to the QTL.}

\item{n_des}{Number of MPP design. Default = 9.}

\item{n_mod}{Number of QTL detection models. Default = 4.}

\item{mod_names}{Models names. Default = c("cr", "par", "anc", "biall").}

\item{MPP_names}{MPP design names.}
}
\value{
list with vector of TPR average per MPP design, and vector of TPR
average per model, matrix of average TPR per MPP and model, sandard deviation
of the TPR per MPP and model, vector of TPR average per MPP design, and
vector of TPR average per model.
}
\description{
Function to organise the TPR results obtained from a list of "true" QTL
positions and a list of detected QTLs in several MPP designs with different
models. The function compute the TPR in each situation (Rep * MPP design *
model). Then it return a table with the average TPR over the x replicate for
each combination of MPP design and QTL detection model. It return a similar
table with the standard deviation. Finally, it also return the global average
per model and per MPP design.
}
\author{
Vincent Garin
}
