% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formMppData2.R
\name{formMppData}
\alias{formMppData}
\title{Form mppData object}
\usage{
formMppData(
  geno_sim,
  nPar,
  map,
  MPP_sub,
  pheno,
  type = NULL,
  F.gen = NULL,
  BC.gen = NULL,
  window = 20
)
}
\arguments{
\item{geno_sim}{genotype matrix of the global population.}

\item{nPar}{Number of parents in the original diallel.}

\item{map}{Correspondig three column genetic map (mk.id, chr, pos cM).}

\item{MPP_sub}{subset information obtained as output of the function
\code{\link{SampleMPPDesign}}.}

\item{pheno}{Simulated phenotype values.}

\item{type}{Type of population.}

\item{F.gen}{Number of F generations.}

\item{BC.gen}{Number of BC generations.}

\item{window}{size of the window used by clusthaplo to cluster the parents.}
}
\value{
mppData object
}
\description{
Function that subset a global population composed from all the cross of a
Dialleld structure and form mppData object for realisation of particular
MPP design.
}
\author{
Vincent Garin
}
