% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPheno.R
\name{SimPheno}
\alias{SimPheno}
\title{Simulation of phenotypic values}
\usage{
SimPheno(QTL, y_poly, k = 1, her, mppData, mppData_bi, par_clu = par_clu)
}
\arguments{
\item{QTL}{list of QTL position obtained with function \code{\link{QposSelect}}}

\item{y_poly}{vector of polygenic contribution.}

\item{k}{Factor representing the ration between QTL (VQ) and polygenic (Vg)
variance (VQ = k * Vg). Default = 1.}

\item{her}{heritability.}

\item{mppData}{IBD \code{mppData} object}

\item{mppData_bi}{IBS \code{mppData} object}

\item{par_clu}{parent clustering object.}
}
\value{
List with the simulated phenotypic values, each contribution
(polygenic, QTL, error), and the proportions of polygenic, QTL, error,
QTL covariance, covariance between QTL and polygenic, etc. And also a list
of extended QTLs effect with the proportion of non-zero QTL effect
contribution (~ QTL allele frequency estimation).
}
\description{
Function simulating phenotypic values based on polygenic, QTL and
(environmental) error component. The simulation admit 8 types of QTLs: Q1
has a different allele in each cross; Q2 segregates in half of the crosses
with a different allelic value; Q3 has a different allele for each parent;
Q4 has a unique allele hold by a single parent; Q5 has a different allele for
each ancestral group; Q6 has a unique allele hold by the most frequent
ancestral group; Q7 and Q8 are bi-allelic QTLs attached to a SNP marker.
}
\author{
Vincent Garin
}
